Function to simulate the average number of leaps, populating a moving avg vector, with tests

+++ src/frog.hpp
@@ -13,4 +13,4 @@
 
 int leap(const int pads_left, const int leaps_made);
 double __declspec(dllexport) average(const int num_pads, const int num_itr);
-void __declspec(dllexport) simulate(const int num_pads, const int num_itr, int * const leaps);
+void __declspec(dllexport) simulate(const int num_pads, const int num_itr, double * const leaps);
